% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtpqroot_xtpqroot.R
\name{print.xtpqroot}
\alias{print.xtpqroot}
\title{Print Method for xtpqroot Objects}
\usage{
\method{print}{xtpqroot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xtpqroot"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for xtpqroot Objects
}
