% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtqsh_xtqsh.R
\name{print.xtqsh}
\alias{print.xtqsh}
\title{Print method for xtqsh objects}
\usage{
\method{print}{xtqsh}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xtqsh"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for xtqsh objects
}
