% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtpcmg_xtpcmg.R
\name{summary.xtpcmg}
\alias{summary.xtpcmg}
\title{Summary Method for xtpcmg Objects}
\usage{
\method{summary}{xtpcmg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xtpcmg"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary Method for xtpcmg Objects
}
