% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtpqroot_xtpqroot.R
\name{summary.xtpqroot}
\alias{summary.xtpqroot}
\title{Summary Method for xtpqroot Objects}
\usage{
\method{summary}{xtpqroot}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xtpqroot"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary Method for xtpqroot Objects
}
