% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtqsh_xtqsh.R
\name{summary.xtqsh}
\alias{summary.xtqsh}
\title{Summary method for xtqsh objects}
\usage{
\method{summary}{xtqsh}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xtqsh"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for xtqsh objects
}
