% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelecon-package.R
\docType{package}
\name{panelecon-package}
\alias{panelecon}
\alias{panelecon-package}
\title{panelecon: Comprehensive Panel Econometric Methods}
\description{
A comprehensive toolkit for panel data econometrics including:
\itemize{
  \item \code{\link{xtpretest}} - Panel pre-testing for unit roots
  \item \code{\link{xtmispanel}} - Missing data detection and imputation
  \item \code{\link{xtcsdq}} - Cross-sectional dependence Q test
  \item \code{\link{xtfifevd}} - Fixed effects vector decomposition
  \item \code{\link{xtpcaus}} - Panel Granger causality tests
  \item \code{\link{xtpcmg}} - Panel cointegrating polynomial regressions
  \item \code{\link{xtqsh}} - Quantile regression slope homogeneity test
  \item \code{\link{xtpqroot}} - Panel quantile unit root tests
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/muhammedalkhalaf}
}

}
\author{
\strong{Maintainer}: Muhammad Alkhalaf \email{muhammedalkhalaf@gmail.com} (\href{https://orcid.org/0009-0002-2677-9246}{ORCID}) [copyright holder]

}
\keyword{internal}
